% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{hydroPck_calculate_idf}
\alias{hydroPck_calculate_idf}
\title{Calculate IDF models}
\usage{
hydroPck_calculate_idf(data)
}
\arguments{
\item{data}{A data.table object containing imported data.}
}
\value{
A data.table containing IDF models.
}
\description{
This function calculates IDF models.
}
\examples{
\dontrun{
  # Assuming `hydroPck_import_nc_files` is correctly implemented and tested
  data <- hydroPck_import_nc_files(path = "/path/to/nc/files")
  hydroPck_calculate_idf(data)
}
}
