% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{hydroPck_download_and_unzip}
\alias{hydroPck_download_and_unzip}
\title{Download and unzip data from a given URL}
\usage{
hydroPck_download_and_unzip(
  url = "https://owncloud.cesnet.cz/index.php/s/HyKD3KXSOontoKX/download"
)
}
\arguments{
\item{url}{Character string. The URL of the zip file to download.}
}
\value{
Invisible(NULL)
}
\description{
Downloads a zip file from the provided URL, extracts it, and prints information about the file.
}
\examples{
\dontrun{
  hydroPck_download_and_unzip()
  custom_url <- "https://example.com/custom_file.zip"
  hydroPck_download_and_unzip(url = custom_url)
}
}
