% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{hydroPck_import_nc_files}
\alias{hydroPck_import_nc_files}
\title{Import NetCDF (*.nc) files}
\usage{
hydroPck_import_nc_files(
  path,
  ids = c(296, 263, 264, 265, 295, 297, 327, 328, 329)
)
}
\arguments{
\item{path}{Character string. The path to the directory containing NetCDF files.}

\item{ids}{Numeric vector. The cell IDs to extract from each NetCDF file. Default is c(296, 263, 264, 265, 295, 297, 327, 328, 329).}
}
\value{
A data.table containing imported data.
}
\description{
This function imports NetCDF (*.nc) files and returns a data.table containing the imported data.
}
\examples{
\dontrun{
  hydroPck_import_nc_files(path = "/path/to/nc/files")
  hydroPck_import_nc_files(path = "/path/to/nc/files", ids = c(1, 2, 3))
}
}
