% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{hydroPck_plot_idf_curves}
\alias{hydroPck_plot_idf_curves}
\title{Plot IDF Curves}
\usage{
hydroPck_plot_idf_curves(idf_data)
}
\arguments{
\item{idf_data}{A data.table containing IDF models.}
}
\value{
A ggplot object displaying IDF curves.
}
\description{
This function plots IDF curves.
}
\examples{
\dontrun{
  # Assuming `hydroPck_calculate_idf` is correctly implemented and tested
  idf_data <- hydroPck_calculate_idf(data)
  hydroPck_plot_idf_curves(idf_data)
}
}
